safari-600.7-branch/LayoutTests/fast/dom/Window/window-frames-self-referential.html

<html>
<head>
<script src="../../../resources/js-test-pre.js"></script>
</head>
<body>

<script>
description('This tests that window.frames is self-referential.');

shouldBe("window.frames", "window");
</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
</html>
