third_party/blink/web_tests/platform/generic/fast/dom/Window/window-frames-self-referential-expected.txt

This tests that window.frames is self-referential.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS window.frames is window
PASS successfullyParsed is true

TEST COMPLETE
